add db command
